+++ src/Endpoint/EndpointInterface.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Onetoweb\Bolcom\Endpoint;
+
+/**
+ * Endpoint Interface.
+ */
+interface EndpointInterface
+{
+    /**
+     * @return string
+     */
+    public static function getName(): string;
+}