+++ src/Endpoint/Inventory.php
@@ -1,0 +1,32 @@
+<?php
+
+namespace Onetoweb\Bolcom\Endpoint;
+
+/**
+ * Inventory Endpoint.
+ * 
+ * @author Jonathan van 't Ende <jvantende@onetoweb.nl>
+ * @copyright Onetoweb B.V.
+ * 
+ * @see https://api.bol.com/retailer/public/redoc/v10/retailer.html#tag/Inventory
+ */
+class Inventory extends AbstractEndpoint
+{
+    /**
+     * {@inheritdoc}
+     */
+    public static function getName(): string
+    {
+        return 'inventory';
+    }
+    
+    /**
+     * @param array $query
+     * 
+     * @return array
+     */
+    public function get(array $query): array
+    {
+        return $this->client->get($this->getUrl(), $query);
+    }
+}